@@ -5,21 +5,21 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.6'
+  sdk: '>=3.2.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   camera: ^0.10.5
-  cupertino_icons: ^1.0.5
-  google_mlkit_face_detection: ^0.6.0
-  flutter_speed_dial: ^6.2.0
+  cupertino_icons: ^1.0.5  
+  flutter_speed_dial: ^7.0.0
   path: ^1.8.2
   path_provider: ^2.0.14
-  permission_handler: ^10.2.0
-  http: ^0.13.5
-  flutter_modular: ^5.0.3
-  device_info_plus: ^8.1.0
+  permission_handler: ^11.1.0
+  dio: ^5.3.2  
+  #http: ^1.1.0
+  flutter_modular: ^6.3.2
+  device_info_plus: ^9.1.1
   bordered_text: ^2.0.0
   badges: ^3.0.3
   icon_badge: ^1.0.0
@@ -33,8 +33,8 @@ dependencies:
   http_parser: ^4.0.2
   intl: ^0.18.0
   image: ^4.0.15
-  tflite_flutter: ^0.9.5
-  #google_ml_kit: ^0.14.0
+  tflite_flutter: ^0.10.3
+  google_ml_kit: ^0.16.3
   #tflite_flutter:
   #  path: assets/tflite_flutter
   ml_linalg: ^13.11.30
@@ -43,8 +43,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: ^0.12.0
-  flutter_lints: ^2.0.1
+  flutter_launcher_icons: ^0.13.1
+  flutter_lints: ^3.0.1
   build_runner: ^2.3.3
 
 
