@@ -1,6 +1,6 @@
 import 'dart:ui';
 import 'package:camera/camera.dart';
-import 'package:google_mlkit_face_detection/google_mlkit_face_detection.dart';
+import 'package:google_ml_kit/google_ml_kit.dart';
 
 class CameraService {
   bool? _isBusyCamera;
