@@ -2,9 +2,9 @@ import 'package:flutter_modular/flutter_modular.dart';
 
 class LoginModule extends Module {
   @override
-  final List<Bind> binds = [];
+  void binds(Injector i) {}
 
   @override
-  final List<ModularRoute> routes = [];
+  void routes(RouteManager r) {}
 
 }
